% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercury.r
\name{just_the_facts}
\alias{just_the_facts}
\title{Retrieve parsed content of a URL processed by the Postlight Mercury API}
\usage{
just_the_facts(url, mercury_api_key = Sys.getenv("MERCURY_API_KEY"))
}
\arguments{
\item{url}{URL to retrieve}

\item{mercury_api_key}{your Mercury API key. The function looks for it in \code{MERCURY_API_KEY}
but you can specify it manually as well. Get your key \href{https://mercury.postlight.com}{here}.}
}
\value{
\code{data.frame}
}
\description{
Mercury takes any web article and returns only the relevant content — headline, author,
body text, relevant images and more — free from any clutter.
}
