% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgr-package.R
\docType{package}
\name{hgr}
\alias{hgr}
\alias{hgr-package}
\title{Tools to Work with the 'Postlight' 'Mercury' 'API'}
\description{
\href{https://mercury.postlight.com}{Mercury} takes any web article and returns only the
relevant content --- headline, author, body text, relevant images and more --- free
from any clutter.
}
\details{
Tools are provided to access the 'API' and also further clean up retrieved text
through the the application of 'XSLT' style sheets. An 'RStudio' 'Addin' is also
provided which makes it possible to preview the cleaned content from a 'URL' on
the clipboard.
}
\author{
Bob Rudis (bob@rud.is)
}
