% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgr-package.R
\docType{package}
\name{hgr}
\alias{hgr}
\alias{hgr-package}
\title{Tools to Work with the 'Postlight' 'Mercury' 'API'}
\description{
Mercury takes any web article and returns only the relevant content — headline, author,
body text, relevant images and more — free from any clutter.
}
\author{
Bob Rudis (bob@rud.is)
}
